src/gateway_service/requirements.txt
fastapi[all]
pydantic
kafka-python
websockets
